change-pass-vm: update to use a token
This new version has sane defaults when used on an engine installed from the ISO
For other scenarios where certificates or other configurations have to be
changed, the -c argument can be used to specify a configuration file.
The --dump-config flag can be used to generate a new configuration that can be
used as a model to create a new one.

For example:

change-vm-pass --dump-config ctx 123 123 > /tmp/my-new-config.yml
modify /tmp/my-new-config.yml
change externpass in the advanced voicemail config to
/usr/share/asterisk/bin/change-pass-vm -c /tmp/my-new-config.yml